+++ client/src/lib/index.ts
@@ -1,5 +1,4 @@
-export * from './conversationsView/index.js';
-export * from './conversation/index.js';
+export * from './conversations/index.js';
 export * from './room/index.js';
 export * from './API.js';
 export * from './types.js';