+++ client/CHANGELOG.md
@@ -1,5 +1,11 @@
 # missiv-client
 
+## 0.0.6
+
+### Patch Changes
+
+- use store for comversations opening state
+
 ## 0.0.5
 
 ### Patch Changes